refactor(backend): extract test run crud module
Move the test run CRUD out of the crud/__init__.py monolith into a dedicated
crud/test_run.py, following crud/token.py and crud/metric.py and the split rule
in apps/backend/AGENTS.md. Pure move: every function body is byte-identical to
the original, verified by parsing both files with ast and comparing the
unparsed functions.

The block is self-contained - nothing else in crud/__init__.py calls into it and
it calls nothing else there. _TEST_RUN_RELATED_FIELDS and the private
_defer_endpoint_last_token helper are used only inside the block and move with
it. joinedload and generate_memorable_name were used only here, so they leave
the import header too.

Callers import the functions directly, per the AGENTS.md rule against reaching
through the crud package root. routers/test_run.py and tasks/execution/run.py
import the submodule as test_run_crud instead: their own handlers are also
called create_test_run, update_test_run and delete_test_run, so a bare function
import would shadow them.

The four tasks modules reached schemas through crud.schemas.TestRunUpdate, which
only resolved because crud/__init__.py happens to import schemas. They now
import rhesis.backend.app.schemas directly, which is what lets the crud import
go away entirely.

Stops before the Test Result CRUD section - _TEST_RESULT_RELATED_FIELDS and the
test_result functions stay in crud/__init__.py for a later PR. The test_run
functions query models.TestResult directly but never call those CRUD functions,
so the split is clean.
