+++ apps/backend/src/rhesis/backend/app/routers/experiments.py
@@ -25,6 +25,7 @@ from rhesis.backend.app.auth.principal import resolve_principal_from_request
 from rhesis.backend.app.auth.rbac import authorize_object, project_id_from_scope
 from rhesis.backend.app.auth.user_utils import require_current_user_or_token
 from rhesis.backend.app.crud.project import get_project
+from rhesis.backend.app.crud.test_run import get_test_runs
 from rhesis.backend.app.dependencies import (
     get_tenant_context,
     get_tenant_db_session,
@@ -380,7 +381,7 @@ def get_experiment_results(
         user_id=user_id,
     )
 
-    runs = crud.get_test_runs(
+    runs = get_test_runs(
         db,
         skip=0,
         limit=limit,
