+++ apps/backend/src/rhesis/backend/tasks/utils.py
@@ -8,7 +8,8 @@ from typing import Any, Dict, Optional, Tuple
 
 from sqlalchemy.orm import Session
 
-from rhesis.backend.app import crud
+from rhesis.backend.app import schemas
+from rhesis.backend.app.crud.test_run import get_test_run, get_test_runs, update_test_run
 from rhesis.backend.app.utils.uuid_utils import safe_uuid_convert
 from rhesis.backend.tasks.enums import RunStatus
 
@@ -30,7 +31,7 @@ def get_test_run_by_config(
         Most recent test run or None if not found
     """
     try:
-        test_runs = crud.get_test_runs(
+        test_runs = get_test_runs(
             db,
             limit=limit,
             filter=f"test_configuration_id eq {test_configuration_id}",
@@ -61,7 +62,7 @@ def get_test_run_by_task_id(
         # Get all test runs and filter by task_id in attributes
         # Note: This is not the most efficient approach, but since we expect
         # few test runs per task, it's acceptable for now
-        test_runs = crud.get_test_runs(db, limit=100, organization_id=organization_id)
+        test_runs = get_test_runs(db, limit=100, organization_id=organization_id)
 
         for test_run in test_runs:
             if test_run.attributes and test_run.attributes.get("task_id") == task_id:
@@ -101,9 +102,7 @@ def increment_test_run_progress(
         if not test_run_uuid:
             return False
 
-        test_run = crud.get_test_run(
-            db, test_run_uuid, organization_id=organization_id, user_id=user_id
-        )
+        test_run = get_test_run(db, test_run_uuid, organization_id=organization_id, user_id=user_id)
         if not test_run:
             return False
 
@@ -134,10 +133,10 @@ def increment_test_run_progress(
         # Update the test run
         update_data = {"attributes": current_attributes}
 
-        crud.update_test_run(
+        update_test_run(
             db,
             test_run.id,
-            crud.schemas.TestRunUpdate(**update_data),
+            schemas.TestRunUpdate(**update_data),
             organization_id=str(test_run.organization_id) if test_run.organization_id else None,
             user_id=str(test_run.user_id) if test_run.user_id else None,
         )
